+++ Server/src/main/java/ForMZ/Server/domain/commentLike/controller/CommentLikeController.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.commentLike.controller;
+
+public class CommentLikeController {
+}
