+++ Server/src/main/java/ForMZ/Server/domain/commentLike/dto/CommentLikeReq.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.commentLike.dto;
+
+public class CommentLikeReq {
+}
