+++ Server/src/main/java/ForMZ/Server/domain/commentLike/exception/CommentLikeNotFoundException.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.commentLike.exception;
+
+public class CommentLikeNotFoundException {
+}
