+++ Server/src/main/java/ForMZ/Server/domain/commentLike/mapper/CommentLikeMapper.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.commentLike.mapper;
+
+public interface CommentLikeMapper {
+}
