+++ Server/src/main/java/ForMZ/Server/domain/commentLike/repository/CommentLikeRepository.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.commentLike.repository;
+
+public interface CommentLikeRepository {
+}
