+++ Server/src/main/java/ForMZ/Server/domain/commentLike/service/CommentLikeService.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.commentLike.service;
+
+public class CommentLikeService {
+}
