--- Server/src/main/java/ForMZ/Server/domain/like/commentLike/controller/CommentLikeController.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.commentLike.controller;
-
-public class CommentLikeController {
-}
