--- Server/src/main/java/ForMZ/Server/domain/like/commentLike/dto/CommentLikeReq.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.commentLike.dto;
-
-public class CommentLikeReq {
-}
