--- Server/src/main/java/ForMZ/Server/domain/like/commentLike/exception/CommentLikeNotFoundException.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.commentLike.exception;
-
-public class CommentLikeNotFoundException {
-}
