--- Server/src/main/java/ForMZ/Server/domain/like/commentLike/mapper/CommentLikeMapper.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.commentLike.mapper;
-
-public interface CommentLikeMapper {
-}
