--- Server/src/main/java/ForMZ/Server/domain/like/commentLike/repository/CommentLikeRepository.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.commentLike.repository;
-
-public interface CommentLikeRepository {
-}
