--- Server/src/main/java/ForMZ/Server/domain/like/commentLike/service/CommentLikeService.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.commentLike.service;
-
-public class CommentLikeService {
-}
